+++ tests/Database/Adapter/PdoAdapterTest.php
@@ -87,21 +87,21 @@ class PdoAdapterTest extends TestCase
 
         $this->adapter->execute('CREATE TABLE "phoenix_test_table" ("id" INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,"title" TEXT NOT NULL,"date" TEXT NOT NULL);');
         $this->assertEquals(1, $this->adapter->insert('phoenix_test_table', ['id' => 1, 'title' => 'first', 'date' => new DateTime()]));
-        $item = $this->adapter->fetch('phoenix_test_table', 'title', ['id' => 1]);
+        $item = $this->adapter->fetch('phoenix_test_table', ['title'], ['id' => 1]);
         $this->assertEquals('first', $item['title']);
 
         $this->assertTrue($this->adapter->update('phoenix_test_table', ['id' => 1, 'title' => 'second', 'date' => new DateTime()], ['id' => 1]));
-        $item = $this->adapter->fetch('phoenix_test_table', 'title', ['id' => 1], ['title']);
+        $item = $this->adapter->fetch('phoenix_test_table', ['title'], ['id' => 1], ['title']);
         $this->assertEquals('second', $item['title']);
 
         $this->assertTrue($this->adapter->update('phoenix_test_table', ['id' => 1, 'title' => 'third', 'date' => new DateTime()], [], 'id = 1'));
-        $items = $this->adapter->fetchAll('phoenix_test_table', 'title', ['id' => 1], null, ['title' => 'DESC'], ['id']);
+        $items = $this->adapter->fetchAll('phoenix_test_table', ['title'], ['id' => 1], null, ['title' => 'DESC'], ['id']);
         $this->assertEquals('third', $items[0]['title']);
 
         $this->assertEquals(2, $this->adapter->insert('phoenix_test_table', ['id' => 2, 'title' => 'fourth', 'date' => new DateTime()]));
         $this->assertTrue($this->adapter->update('phoenix_test_table', ['title' => 'multi update'], ['id' => [1, 2]]));
 
-        $items = $this->adapter->fetchAll('phoenix_test_table', 'title', ['id' => [1, 2]]);
+        $items = $this->adapter->fetchAll('phoenix_test_table', ['title'], ['id' => [1, 2]]);
         $this->assertCount(2, $items);
         foreach ($items as $item) {
             $this->assertEquals('multi update', $item['title']);